update ex5.1 not float

+++ Exercise5_1_yongyut_w.py
@@ -3,4 +3,4 @@ y = 2
 print(x,"+",y,"=",x + y)
 print(x,"-",y,"=",x - y)
 print(x,"*",y,"=",x * y)
-print(x,"/",y,"=",x / y)
+print(x,"/",y,"=",int(x / y))
